+++ src/es/ubu/inf/tfg/asientosContables/CompraPropiedadIndustrialaAmortizable.java
@@ -7,7 +7,7 @@ import es.ubu.inf.tfg.otrasCosas.Enunciado;
 
 public class CompraPropiedadIndustrialaAmortizable extends Asiento{
 	
-	public CompraPropiedadIndustrialaAmortizable(Calendar f, int[] i) {
+	public CompraPropiedadIndustrialaAmortizable(Calendar f, double[] i) {
 		fecha =f;
 		inputs=i;
 
@@ -25,7 +25,7 @@ public class CompraPropiedadIndustrialaAmortizable extends Asiento{
 		
 		//SE SALDAN LAS DEUDAS CON LOS PROVEEDORES "Y" DIAS DESPUES:
 		Calendar fechaDeudas = (Calendar)fecha.clone();
-		fechaDeudas.add(Calendar.DAY_OF_YEAR, +inputs[1]);
+		fechaDeudas.add(Calendar.DAY_OF_YEAR, (int) +inputs[1]);
 		
 		String enunciado2 = " Se salda la deuda con los proveedores de la propiedad industrial.\n"
 				+ "CUENTAS PGC: 400. Proveedores; 572. Bancos e instituciones de crédito c/c vista, euros. \n";
